cli: helpers
